Добавил возможность передавать глюкозу и углеводы в единицах измерения, настроенных больным, не только для записей дневника, но и для профиля пациента (пороговые значения) и запроса на расчет инсулина; + теперь валидация диапазонов значений глюкоз и углеводов зависит от единицы измерения;

+++ src/main/java/com/artlighter/glucosecontrolservice/templates/util/mapper/MealMapper.java
@@ -16,13 +16,16 @@ public class MealMapper implements TemplateMapper<Meal, MealDTO> {
 
     @Override
     public Meal mapToInternal(MealDTO externalDTO) {
+        //TODO решить, стоит ли принимать кол-во углеводов на 100 г тоже в единицах пользователя или только в граммах.
+        //Просто может быть удобнее на входе передавать в граммах независимо от выставленных единиц, ведь именно граммы
+        //написаны на упаковках. Хотя еда необязательно упаковочная...
         return new Meal(new PatientTemplateEntity.PatientTemplateEntityID(0, externalDTO.name()),
                 externalDTO.carbsPer100Grams());
     }
 
     public MealDTO mapToDtoWithUnitConversion(Meal internal, CarbsUnit patientCarbsUnit) {
         return new MealDTO(internal.getId().getName(),
-                patientCarbsUnit.convertFromGrams(internal.getCarbsPer100Grams()));
+                (float) patientCarbsUnit.convertFromGrams(internal.getCarbsPer100Grams()));
     }
 
 }
